+++ addons/godotsteam_csharpbindings/Steam.Video.cs
@@ -2,17 +2,17 @@ namespace GodotSteam;
 
 public static partial class Steam
 {
-    public static void GetOPFSettings(long appId)
+    public static void GetOPFSettings(uint appId)
     {
         GetInstance().Call(Methods.GetOPFSettings, appId);
     }
     
-    public static string GetOPFStringForApp(long appId)
+    public static string GetOPFStringForApp(uint appId)
     {
         return GetInstance().Call(Methods.GetOPFStringForApp, appId).AsString();
     }
     
-    public static void GetVideoURL(long appId)
+    public static void GetVideoURL(uint appId)
     {
         GetInstance().Call(Methods.GetVideoURL, appId);
     }
